+++ tradingagents/agents/__init__.py
@@ -17,7 +17,7 @@ from .risk_mgmt.conservative_debator import create_conservative_debator
 from .risk_mgmt.neutral_debator import create_neutral_debator
 
 from .managers.research_manager import create_research_manager
-from .managers.portfolio_manager import create_portfolio_manager
+from .managers.asset_manager import create_asset_manager
 
 from .trader.trader import create_trader
 
@@ -34,7 +34,7 @@ __all__ = [
     "create_neutral_debator",
     "create_news_analyst",
     "create_aggressive_debator",
-    "create_portfolio_manager",
+    "create_asset_manager",
     "create_conservative_debator",
     "create_sentiment_analyst",
     "create_social_media_analyst",  # deprecated; will be removed in a future version